+++ cli/src/main.rs
@@ -19,7 +19,7 @@ use structopt::{clap::Shell as ClapShell, StructOpt};
 
 use crate::{
     args::{Args, Command, Shell},
-    commands::{config as config_command, create, delete, get, update},
+    commands::{config as config_command, create, delete, get, parse, update},
     config::ReinferConfig,
     printer::Printer,
 };
@@ -71,6 +71,7 @@ fn run(args: Args) -> Result<()> {
         Command::Update { update_args } => {
             update::run(update_args, client_from_args(&args, &config)?, &printer)
         }
+        Command::Parse { parse_args } => parse::run(parse_args, client_from_args(&args, &config)?),
     }
 }